+++ cs3560_grading_lib/common.py
@@ -81,7 +81,7 @@ def run_executable(args, cwd: Optional[str | Path] = None) -> CommandResult:
     """
     try:
         make_cmd_output = subprocess.check_output(
-            args, stderr=subprocess.STDOUT, cwd=cwd
+            args, stdout=subprocess.STDOUT, stderr=subprocess.STDOUT, cwd=cwd
         )
         return CommandResult(True, " ".join(args), make_cmd_output.decode())
     except subprocess.CalledProcessError as e: